6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,7 +11,7 @@ int main(void)
 	char a;
 	int b;
 	long int c;
-	long long int d;
+	long int d;
 	float f;
 
 printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
